refactor: extract deployment logic to new file
